+++ src/main/java/com/project/service/UserService.java
@@ -1,16 +1,17 @@
 package com.project.service;
 
-import com.project.domain.user.User;
+import com.project.domain.User;
 
 import java.util.List;
 
 public interface UserService {
+
     User login(String email, String password);
 
     boolean register(User user);
 
     List<User> showAll(Integer startFrom, Integer rowCount);
 
-    Integer showEntriesAmount();
+    Long showEntriesAmount();
 
 }
